refactor: Centralise mapToStyleProperty and utils
