+++ lib/atoms/utils/rowSpacingForCssRule.ts
@@ -1,7 +1,7 @@
 import toPairs from 'lodash/toPairs';
 import fromPairs from 'lodash/fromPairs';
 
-import { px } from './toUnit';
+import { px } from '../../utils/toUnit';
 import { Tokens } from '../../themes/theme';
 
 export default (
