+++ lib/hooks/useBox/box.treat.ts
@@ -1,22 +1,10 @@
-import mapValues from 'lodash/mapValues';
 import { css, Styles, style } from 'sku/treat';
 import { Properties } from 'csstype';
 import { darken, lighten } from 'polished';
-import { getLightVariant } from '../../atoms/utils/a11y';
-import isLight from '../../atoms/utils/isLight';
-import { px } from '../../atoms/utils/toUnit';
-
-const mapToCssRule = <Map extends string, Value extends string | number>(
-  map: Record<Map, Value>,
-  propertyName: keyof Properties,
-  mapper?: (value: Value, propertyName: keyof Properties) => Styles,
-) => {
-  type MappedStyles = Record<Map, Styles>;
-
-  return mapValues(map, (value: Value) =>
-    mapper ? mapper(value, propertyName) : { [propertyName]: value },
-  ) as MappedStyles;
-};
+import { getLightVariant } from '../../utils/a11y';
+import isLight from '../../utils/isLight';
+import { px } from '../../utils/toUnit';
+import mapToStyleProperty from '../../utils/mapToStyleProperty';
 
 const spaceMapToCss = <Map extends string>(
   spaceMap: Record<Map, number>,
@@ -29,7 +17,7 @@ const spaceMapToCss = <Map extends string>(
     none: 0,
   };
 
-  return mapToCssRule(
+  return mapToStyleProperty(
     spaceMapWithNone,
     cssPropertyName,
     (value: number, propertyName) =>
@@ -80,17 +68,17 @@ export const transform = {
 };
 
 export const transition = css(({ transitions }) =>
-  mapToCssRule(transitions, 'transition'),
+  mapToStyleProperty(transitions, 'transition'),
 );
 
 export const borderRadius = css(({ border }) =>
-  mapToCssRule(border.radius, 'borderRadius'),
+  mapToStyleProperty(border.radius, 'borderRadius'),
 );
 
 const widthRules = {
   full: '100%',
 };
-export const width = css(mapToCssRule(widthRules, 'width'));
+export const width = css(mapToStyleProperty(widthRules, 'width'));
 
 const displayRules = {
   block: 'block',
@@ -99,9 +87,9 @@ const displayRules = {
   inlineBlock: 'inline-block',
   flex: 'flex',
 };
-export const display = css(mapToCssRule(displayRules, 'display'));
+export const display = css(mapToStyleProperty(displayRules, 'display'));
 export const displayDesktop = css(({ utils: { desktopStyles } }) =>
-  mapToCssRule(displayRules, 'display', (value, propertyName) =>
+  mapToStyleProperty(displayRules, 'display', (value, propertyName) =>
     desktopStyles({
       [propertyName]: value,
     }),
@@ -113,13 +101,16 @@ const flexDirectionRules = {
   column: 'column',
 };
 export const flexDirection = css(
-  mapToCssRule(flexDirectionRules, 'flexDirection'),
+  mapToStyleProperty(flexDirectionRules, 'flexDirection'),
 );
 export const flexDirectionDesktop = css(({ utils: { desktopStyles } }) =>
-  mapToCssRule(flexDirectionRules, 'flexDirection', (value, propertyName) =>
-    desktopStyles({
-      [propertyName]: value,
-    }),
+  mapToStyleProperty(
+    flexDirectionRules,
+    'flexDirection',
+    (value, propertyName) =>
+      desktopStyles({
+        [propertyName]: value,
+      }),
   ),
 );
 
@@ -130,7 +121,7 @@ const getHoverColor = (color: string) =>
   isLight(color) ? darken(0.05, color) : lighten(0.05, color);
 
 export const backgroundColor = css(({ color: { background } }) => ({
-  ...mapToCssRule(background, 'backgroundColor'),
+  ...mapToStyleProperty(background, 'backgroundColor'),
   formAccentActive: { backgroundColor: getActiveColor(background.formAccent) },
   formAccentHover: { backgroundColor: getHoverColor(background.formAccent) },
   brandAccentActive: {
